Change the way points are treated in models
Use a Point2d object to represent a single (x, y) pair instead of using one array for x values and another for y values. Also make it so that plots will scale arbitrary values to fit into them instead of depending on the x and y values to match up with the pixels numbers in the canvas.